Drop open telemetry
