+++ racetrack_commons/requirements.txt
@@ -3,8 +3,6 @@ pydantic==2.6.1
 pyjwt[crypto]==2.6.0
 fastapi==0.110.1
 uvicorn[standard]==0.29.0
-opentelemetry-exporter-otlp-proto-http==1.24.0
-protobuf==4.25.3  # needed by opentelemetry exporter
 python-multipart==0.0.9  # uploading files
 watchdog==2.3.1
 a2wsgi==1.10.4
